Refactor: to Init DDD (Domain Driven Design)

+++ linker-backend/src/main/kotlin/linker/service/BoardService.kt
@@ -1,14 +1,14 @@
 package linker.service
 
 import linker.dto.BoardDto
-import linker.dto.CategoryDto
-import linker.dto.LinkDto
 import linker.dto.UpdateBoardCommand
+import linker.dto.UserDto
 import linker.entity.Board
 import linker.entity.Category
 import linker.repository.BoardRepository
 import linker.repository.CategoryRepository
 import linker.repository.LinkRepository
+import linker.repository.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -30,6 +30,8 @@ class BoardService {
     @Autowired
     lateinit var linkRepository: LinkRepository
     @Autowired
+    lateinit var userRepository: UserRepository
+    @Autowired
     lateinit var userService: UserService
 
     fun findBoardById(id: Long): Board {
@@ -39,9 +41,7 @@ class BoardService {
     @Transactional
     fun getBoard(id: Long): BoardDto =
         findBoardById(id = id).let { board->
-            BoardDto.fromDomain(board = board, categories = board.categories.map { category ->
-                CategoryDto.fromDomain(category = category, links = category.links.map { link ->
-                    LinkDto.fromDomain(link = link) }.sortedBy { it.order }) }.sortedBy { it.order })
+            BoardDto.fromDomain(board = board, user = UserDto.fromDomain(user = userRepository.findById(board.userId ?: 0).orElse(null)))
         }
 
     @Transactional
@@ -49,11 +49,11 @@ class BoardService {
         val user = userService.findByEmail(updateBoardCommand.email)
         val categories = categoryService.findAll().map { category -> category.id }
         categories.subtract(updateBoardCommand.categories.map { categoryDto -> categoryDto.id }).forEach{id ->
-            categoryRepository.deleteById(id)
+            if (id!= null) categoryRepository.deleteById(id)
         }
         val board = this.findBoardById(id)
         updateBoardCommand.categories.forEachIndexed { index, categoryDto ->
-            val category = categoryRepository.findById(categoryDto.id).orElse(Category(0, "", 0, board, emptyList()))
+            val category = categoryRepository.findById(categoryDto.id!!).orElse(Category(0, "", 0, board.id ?: 0, emptyList()))
             val links = categoryDto.links?.mapIndexed { linkIndex, linkDto ->
                 val link = linkDto.toDomain(category = category, user = user)
                 link.order = linkIndex+1
@@ -62,7 +62,7 @@ class BoardService {
             } ?: emptyList()
             category.links = links
             category.title = categoryDto.title
-            category.id = categoryDto.id
+            category.id = categoryDto.id!!
             category.order = index+1
             categoryRepository.save(category)
         }
